Time: 1 ms (100%), Space: 46.4 MB (27.02%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.java
@@ -27,7 +27,7 @@ class Solution {
 
        
         if (tl % 2 != 0) { 
-    return (double) arr[tl / 2];
+    return arr[tl / 2];
 } else { 
     return (arr[tl/2 - 1] + arr[tl/2]) / 2.0;
 }
